backend/requirements.txt
PyPDF2
google-genai
python-dotenv